refactor: Move types to bridge package
This allows us to define these types at a lower level and have a one-way dependency on the bridge types from 'js-miniapp-sdk'. This also lets us use the 'MiniAppBridge' type in the 'js-miniapp-sdk' code to ensure that the SDK is conforming to what the bridge code expects
