+++ js-miniapp-bridge/src/index.ts
@@ -1,0 +1,29 @@
+/** @internal */
+
+/**
+ * Exports the types for Mini App Bridge
+ */
+
+import { MiniAppBridge } from './common-bridge';
+import { AdTypes } from './types/ad-types';
+import { Reward } from './types/response-types/rewarded';
+import { DevicePermission } from './types/device-permission';
+import {
+  CustomPermission,
+  CustomPermissionName,
+  CustomPermissionStatus,
+  CustomPermissionResult,
+} from './types/custom-permissions';
+import { ShareInfoType } from './types/share-info';
+
+export {
+  MiniAppBridge,
+  AdTypes,
+  Reward,
+  DevicePermission,
+  CustomPermission,
+  CustomPermissionName,
+  CustomPermissionStatus,
+  CustomPermissionResult,
+  ShareInfoType,
+};
