+++ js-miniapp-bridge/src/types/ad-types/index.ts
