+++ js-miniapp-bridge/src/types/device-permission.ts
