+++ js-miniapp-bridge/src/types/response-types/index.ts
