+++ js-miniapp-bridge/src/types/response-types/rewarded/index.ts
@@ -1,5 +1,3 @@
-import { AdTypes } from '../../adTypes';
-
 /**
  * A contract declaring the reward response, consists of amount and type of the reward
  */
