+++ js-miniapp-bridge/src/types/share-info.ts
