+++ js-miniapp-sdk/src/index.ts
@@ -4,23 +4,22 @@
  * Main entry point for SDK
  */
 
-import { MiniApp } from './miniapp';
-import { AdTypes } from './types/adTypes';
-import { Reward } from './types/responseTypes/rewarded';
-import { ShareInfoType } from './types/ShareInfoType';
 import {
+  Reward,
   CustomPermission,
   CustomPermissionName,
   CustomPermissionStatus,
   CustomPermissionResult,
-} from './types/CustomPermission';
+  ShareInfoType,
+} from '../../js-miniapp-bridge/src';
+
+import { MiniApp } from './miniapp';
 
 /** @internal */
 const miniAppInstance = new MiniApp();
 
 export default miniAppInstance;
 export {
-  AdTypes,
   CustomPermission,
   CustomPermissionName,
   CustomPermissionStatus,
